ManagerDependencyInjection
Abstracted the View and Controller classes into View Manager and ControllerManager
