+++ managers/ViewManager.py
@@ -1,4 +1,14 @@
 """The purpose of this manager is to intantiate all the view objects and supply it to the user of the objects
 """
+from controller.RecordingController import RecordingController
+from controller.VideoController import VideoController
+from view import RecordingView
+from view.AnnotationEditView import AnnotationEditView
+from view import *
 class ViewManager:
-    pass
+
+    def getAnnotationView(self, recordingController: RecordingController): 
+        return AnnotationEditView(recordingController)
+        
+    def getVideoView(self, videoController: VideoController):
+        return VideoView(videoController)
